delete method throws Exception NotFoundException

+++ kpoint-back/src/main/java/ua/in/kp/service/SuggestionService.java
@@ -11,6 +11,7 @@ import ua.in.kp.dto.suggestion.SuggestionResponseDto;
 import ua.in.kp.entity.LikeEntity;
 import ua.in.kp.entity.SuggestionEntity;
 import ua.in.kp.entity.UserEntity;
+import ua.in.kp.exception.NotFoundException;
 import ua.in.kp.mapper.SuggestionMapper;
 import ua.in.kp.repository.LikeRepository;
 import ua.in.kp.repository.SuggestionRepository;
@@ -66,6 +67,9 @@ public class SuggestionService {
     }
 
     public void deleteSuggestion(String suggestionId) {
+        if (!suggestionRepository.existsById(suggestionId)) {
+            throw new NotFoundException("Suggestion not found with ID: " + suggestionId);
+        }
         suggestionRepository.deleteById(suggestionId);
     }
 }
